requirements.txt
customtkinter>=5.2.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyinstaller>=6.0.0
